requirements.txt
google_images_download==2.5.0
discord.py==1.2.5
gTTS==2.0.3
